les petits bugs de calcul corriges

+++ repo/src/init/init_player.c
@@ -6,7 +6,7 @@
 /*   By: qhauuy <qhauuy@student.42mulhouse.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 15:17:43 by qhauuy            #+#    #+#             */
-/*   Updated: 2024/10/24 14:24:51 by qhauuy           ###   ########.fr       */
+/*   Updated: 2024/10/24 15:42:09 by qhauuy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@ void	init_player(t_data *data)
 	fill_image(data->player, 0xFF0000FF);
 	if (mlx_image_to_window(data->mlx, data->player, COL_START * SIZE_BOX, ROW_START * SIZE_BOX) == -1)
 		return (mlx_delete_image(data->mlx, data->player), mlx_perror_exit(data));
-	data->x = data->player->instances[0].x + SIZE_PLAYER / 2;
-	data->y = data->player->instances[0].y + SIZE_PLAYER / 2;
+	data->x = data->player->instances[0].x + (double)SIZE_PLAYER / 2;
+	data->y = data->player->instances[0].y + (double)SIZE_PLAYER / 2;
 }
